[Clang] Introduce Swift async calling convention.
This change is intended as initial setup. The plan is to add
more semantic checks later. Most of the code closely mirrors
that for the Swift calling convention. Three places are
marked as [FIXME: swiftasynccc]; those will be addressed in
follow-up commits.

Differential Revision: https://reviews.llvm.org/D95561

(cherry-pick from not-yet-merged revision to apple/stable)

This is a revised version of ad8c8b15b8c1aedd938290407ca0d5f22c9c66b0.

Changes:
- Removed TargetInfo::isSwiftAsyncCCSupported API.